requirements.txt
requests==2.28.1
numpy==1.22.4
pandas==1.4.2
Pydap==3.2.2
SQLAlchemy==1.4.37
PyMySQL==1.0.2
joblib == 1.1.0
sklearn == 0.0
PyYAML == 6.0
rpy2 == 3.5.4
configparser == 5.3.0